Add API module with GraphQL helpers
